refactoring registration hub
