+++ src/components/LoginButton.tsx
@@ -2,8 +2,10 @@ import { Button, Popover, Spin } from "antd";
 import * as wallet from "../services/wallets/wallet";
 import React from "react";
 import { userUpdateWalletType } from "../redux/slices/userSlice";
-import { useAppDispatch } from "../redux/hooks";
+import { useAppDispatch, useAppSelector } from "../redux/hooks";
 import { upsertSessionWalletType } from "../services/session";
+import UserAvatar from "./UserAvatar";
+import { HexString, Profile } from "../utilities/types";
 
 interface LoginButtonProps {
   popoverVisible: boolean;
@@ -20,6 +22,13 @@ const LoginButton = ({
   loading,
   loginWithWalletType,
 }: LoginButtonProps): JSX.Element => {
+  const userId = useAppSelector((state) => state.user.id);
+  const currentWalletType = useAppSelector((state) => state.user.walletType);
+  const profiles: Record<HexString, Profile> = useAppSelector(
+    (state) => state.profiles.profiles
+  );
+  const profile: Profile | undefined = userId ? profiles[userId] : undefined;
+
   const handleVisibleChange = (visible: boolean) => {
     setPopoverVisible(visible);
   };
@@ -58,10 +67,25 @@ const LoginButton = ({
         </div>
       }
     >
-      <Button className="LoginButton__loginButton" aria-label="Login">
-        Log In
-        {loading && <Spin className="LoginButton__spinner" size="small" />}
-      </Button>
+      {!userId ? (
+        <Button className="LoginButton__loginButton" aria-label="Login">
+          Log In
+          {loading && <Spin className="LoginButton__spinner" size="small" />}
+        </Button>
+      ) : (
+        <div className="RegistrationHub__userBlock">
+          <UserAvatar
+            icon={profile?.icon?.[0]?.href}
+            profileAddress={userId}
+            avatarSize="small"
+          />
+          <div className="RegistrationHub__userTitle">
+            {profile?.handle
+              ? "@" + profile.handle
+              : profile?.name || profile?.fromId}
+          </div>
+        </div>
+      )}
     </Popover>
   );
